+++ src/pages/SharedLayout.js
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Navbar from '../components/Navbar';
+import Navbar from '../components/Navbar/Navbar';
 
 const SharedLayout = () => (
   <main>
